refactor: Move identity_verification_service.dart to services folder
